Redoing all of the tests and splitting up helpers and handlers
