+++ pages/api/account/checkouts/renew/renewCheckout.ts
@@ -1,0 +1,15 @@
+import sierraClient from "../../../../../src/server/sierraClient"
+export default async function renewCheckout(checkoutId: string) {
+  try {
+    const client = await sierraClient()
+    const response = await client.post(
+      `patrons/checkouts/${checkoutId}/renewal`
+    )
+    return { status: 200, message: "Renewed", body: response }
+  } catch (error) {
+    return {
+      status: error.response.status,
+      message: error.response.data.message || error.response.data.description,
+    }
+  }
+}
