+++ pages/api/account/holds/update/updateHold.ts
@@ -1,0 +1,14 @@
+import sierraClient from "../../../../../src/server/sierraClient"
+
+export default async function updateHold(holdId: string, holdData: any) {
+  try {
+    const client = await sierraClient()
+    await client.put(`patrons/holds/${holdId}`, holdData)
+    return { status: 200, message: "Updated" }
+  } catch (error) {
+    return {
+      status: error.response.status,
+      message: error.response.data.message || error.response.data.description,
+    }
+  }
+}
