+++ pages/api/account/settings/[id].ts
@@ -1,6 +1,6 @@
 import type { NextApiResponse, NextApiRequest } from "next"
-import sierraClient from "../../../../src/server/sierraClient"
 import initializePatronTokenAuth from "../../../../src/server/auth"
+import updateSettings from "./updateSettings"
 
 /**
  * API route handler for /api/account/settings/{patronId}
@@ -29,7 +29,7 @@ export default async function handler(
     /**  We check that the patron cookie matches the patron id in the request,
      * i.e.,the logged in user is updating their own settings. */
     if (patronId == cookiePatronId) {
-      const response = await settingsUpdate(patronId, patronData)
+      const response = await updateSettings(patronId, patronData)
       responseStatus = response.status
       responseMessage = response.message
     } else {
@@ -38,17 +38,4 @@ export default async function handler(
     }
   }
   res.status(responseStatus).json(responseMessage)
-}
-
-export async function settingsUpdate(patronId: string, patronData: any) {
-  try {
-    const client = await sierraClient()
-    await client.put(`patrons/${patronId}`, patronData)
-    return { status: 200, message: "Updated" }
-  } catch (error) {
-    return {
-      status: error.response.status,
-      message: error.response.data.message || error.response.data.description,
-    }
-  }
 }
