+++ pages/api/account/update-pin/updatePin.ts
@@ -1,0 +1,23 @@
+import sierraClient from "../../../../src/server/sierraClient"
+
+export default async function updatePin(
+  patronId: string,
+  barcode: string,
+  oldPin: string,
+  newPin: string
+) {
+  try {
+    const client = await sierraClient()
+    await client.post("patrons/validate", {
+      barcode: barcode,
+      pin: oldPin,
+    })
+    await client.put(`patrons/${patronId}`, { pin: newPin })
+    return { status: 200, message: `Pin updated to ${newPin}` }
+  } catch (error) {
+    return {
+      status: error.response.status,
+      message: error.response.data.message || error.response.data.description,
+    }
+  }
+}
